Moved proxy_pointer to common file and made it explicit
