+++ projects/rocprim/rocprim/include/rocprim/iterator/arg_index_iterator.hpp
@@ -28,6 +28,7 @@
 
 #include "../config.hpp"
 #include "../types/key_value_pair.hpp"
+#include "detail/common.hpp"
 
 /// \addtogroup iteratormodule
 /// @{
@@ -58,29 +59,6 @@ class arg_index_iterator
 private:
     using input_category = typename std::iterator_traits<InputIterator>::iterator_category;
 
-    template<typename T>
-    class proxy_pointer
-    {
-        T value_;
-
-    public:
-        ROCPRIM_HOST_DEVICE inline proxy_pointer(const T& value) : value_(value) {}
-
-        ROCPRIM_HOST_DEVICE
-        const T*
-            operator->() const
-        {
-            return &value_;
-        }
-
-        ROCPRIM_HOST_DEVICE
-        const T&
-            operator*() const
-        {
-            return value_;
-        }
-    };
-
 public:
     /// The type of the value that can be obtained by dereferencing the iterator.
     using value_type = ::rocprim::key_value_pair<Difference, InputValueType>;
@@ -89,7 +67,7 @@ public:
     using reference = const value_type&;
     /// \brief A pointer type of the type iterated over (\p value_type).
     /// It's `const` since arg_index_iterator is a read-only iterator.
-    using pointer = const proxy_pointer<InputValueType>;
+    using pointer = const detail::proxy_pointer<InputValueType>;
     /// A type used for identify distance between iterators.
     using difference_type = Difference;
     /// The category of the iterator.
